+++ docs/shortest-path_floyd-warshall.js.html
@@ -1,0 +1,130 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>JSDoc: Source: shortest-path/floyd-warshall.js</title>
+
+    <script src="scripts/prettify/prettify.js"> </script>
+    <script src="scripts/prettify/lang-css.js"> </script>
+    <!--[if lt IE 9]>
+      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+    <link type="text/css" rel="stylesheet" href="styles/prettify-tomorrow.css">
+    <link type="text/css" rel="stylesheet" href="styles/jsdoc-default.css">
+</head>
+
+<body>
+
+<div id="main">
+
+    <h1 class="page-title">Source: shortest-path/floyd-warshall.js</h1>
+
+    
+
+
+
+    
+    <section>
+        <article>
+            <pre class="prettyprint source linenums"><code>(function (exports) {
+  'use strict';
+
+  var floydWarshall = (function () {
+
+    /**
+     * Matrix used for the algorithm.
+     */
+    var dist;
+
+    /**
+     * Initialize the distance matrix.
+     *
+     * @private
+     * @param {array} graph Distance matrix of the array.
+     * @return {array} Distance matrix used for the algorithm.
+     */
+    function init(graph) {
+      var dist = [];
+      var size = graph.length;
+      for (var i = 0; i &lt; size; i += 1) {
+        dist[i] = [];
+        for (var j = 0; j &lt; size; j += 1) {
+          if (i === j) {
+            dist[i][j] = 0;
+          } else if (isNaN(graph[i][j])) {
+            dist[i][j] = Infinity;
+          } else {
+            dist[i][j] = graph[i][j];
+          }
+        }
+      }
+      return dist;
+    }
+
+    /**
+     * Floyd-Warshall algorithm. Finds the shortest path between each two vertices.&lt;br>&lt;br>
+     * Complexity: O(|V|^3) where V is the number of vertices.
+     *
+     * @public
+     * @module graphs/shortest-path/floyd-warshall
+     * @param {Array} graph A distance matrix of the graph.
+     * @return {Array} Array which contains the shortest distance between each two vertices.
+     * 
+     * @example
+     * var floydWarshall = require('path-to-algorithms/src/graphs/shortest-path/floyd-warshall').floydWarshall;
+     * var distMatrix = [[NaN, 7,   9,   NaN, NaN, 16],
+     *                   [7,   NaN, 10,  15,  NaN, NaN],
+     *                   [9,   10,  NaN, 11,  NaN, 2],
+     *                   [NaN, 15,  11,  NaN, 6,   NaN],
+     *                   [NaN, NaN, NaN, 6,   NaN, 9],
+     *                   [16,  NaN, 2,   NaN, 9,   NaN]];
+     *
+     * // [ [ 0, 7, 9, 20, 20, 11 ],
+     * //   [ 7, 0, 10, 15, 21, 12 ],
+     * //   [ 9, 10, 0, 11, 11, 2 ],
+     * //   [ 20, 15, 11, 0, 6, 13 ],
+     * //   [ 20, 21, 11, 6, 0, 9 ],
+     * //   [ 11, 12, 2, 13, 9, 0 ] ]
+     * var shortestDists = floydWarshall(distMatrix);
+     */
+    return function (graph) {
+      dist = init(graph);
+      var size = graph.length;
+      for (var k = 0; k &lt; size; k += 1) {
+        for (var i = 0; i &lt; size; i += 1) {
+          for (var j = 0; j &lt; size; j += 1) {
+            if (dist[i][j] > dist[i][k] + dist[k][j]) {
+              dist[i][j] = dist[i][k] + dist[k][j];
+            }
+          }
+        }
+      }
+      return dist;
+    };
+  }());
+
+  exports.floydWarshall = floydWarshall;
+
+})(typeof window === 'undefined' ? module.exports : window);</code></pre>
+        </article>
+    </section>
+
+
+
+
+</div>
+
+<nav>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-graphs_others_topological-sort.html">graphs/others/topological-sort</a></li><li><a href="module-graphs_searching_bfs.html">graphs/searching/bfs</a></li><li><a href="module-graphs_searching_dfs.html">graphs/searching/dfs</a></li><li><a href="module-graphs_shortest-path_bellman-ford.html">graphs/shortest-path/bellman-ford</a></li><li><a href="module-graphs_shortest-path_dijkstra.html">graphs/shortest-path/dijkstra</a></li><li><a href="module-graphs_shortest-path_floyd-warshall.html">graphs/shortest-path/floyd-warshall</a></li></ul>
+</nav>
+
+<br class="clear">
+
+<footer>
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-alpha13</a> on Fri Jan 09 2015 14:21:18 GMT+0200 (EET)
+</footer>
+
+<script> prettyPrint(); </script>
+<script src="scripts/linenumber.js"> </script>
+</body>
+</html>
